major refactor to completely isolate legacy gui and napari plugin
