+++ plantseg/io/__init__.py
@@ -1,3 +1,3 @@
-from plantseg.io.io import smart_load, create_tiff, create_h5, load_shape
-from plantseg.io.io import read_tiff_voxel_size, read_h5_voxel_size
 from plantseg.io.io import allowed_data_format, H5_EXTENSIONS, TIFF_EXTENSIONS
+from plantseg.io.io import read_tiff_voxel_size, read_h5_voxel_size
+from plantseg.io.io import smart_load, create_tiff, create_h5, load_shape
