+++ plantseg/legacy_gui/gui_widgets.py
@@ -1,6 +1,7 @@
 import tkinter
 
-from plantseg.legacy_gui import convert_rgb, list_models
+from plantseg.legacy_gui import convert_rgb
+from plantseg.utils import list_models
 from plantseg.legacy_gui import stick_all, stick_new
 from plantseg.legacy_gui.gui_tools import ListEntry, SimpleEntry, FilterEntry, RescaleEntry, MenuEntry, SliderEntry
 from plantseg.predictions.utils import STRIDE_DRAFT, STRIDE_BALANCED, STRIDE_ACCURATE
