+++ plantseg/pipeline/config_validation.py
@@ -5,11 +5,11 @@ import numpy as np
 import torch
 import yaml
 
-from plantseg.legacy_gui import list_models
 from plantseg.pipeline import gui_logger
 from plantseg.pipeline import raw2seg_config_template
-from plantseg.predictions.utils import STRIDE_ACCURATE, STRIDE_BALANCED, STRIDE_DRAFT, get_stride_shape, check_models
+from plantseg.predictions.utils import STRIDE_ACCURATE, STRIDE_BALANCED, STRIDE_DRAFT, get_stride_shape
 from plantseg.segmentation.utils import SUPPORTED_ALGORITMS
+from plantseg.utils import list_models, check_models
 
 deprecated_keys = {'param': 'filter_param'}
 special_keys = {'nuclei_predictions_path', 'is_segmentation'}
