+++ plantseg/viewer/containers.py
@@ -1,19 +1,20 @@
+import webbrowser
+
+from PyQt5.QtCore import Qt
 from magicgui.widgets import MainWindow
 
-from plantseg.viewer.widget.dataprocessing import widget_rescaling, widget_gaussian_smoothing
 from plantseg.viewer.widget.dataprocessing import widget_cropping, widget_add_layers
+from plantseg.viewer.widget.dataprocessing import widget_label_processing
+from plantseg.viewer.widget.dataprocessing import widget_rescaling, widget_gaussian_smoothing
 from plantseg.viewer.widget.io import open_file, export_stacks
-from plantseg.viewer.widget.predictions import widget_unet_predictions, widget_test_all_unet_predictions
 from plantseg.viewer.widget.predictions import widget_iterative_unet_predictions, widget_add_custom_model
+from plantseg.viewer.widget.predictions import widget_unet_predictions, widget_test_all_unet_predictions
+from plantseg.viewer.widget.proofreading.proofreading import widget_clean_scribble
+from plantseg.viewer.widget.proofreading.proofreading import widget_split_and_merge_from_scribbles
 from plantseg.viewer.widget.segmentation import widget_dt_ws, widget_agglomeration
-from plantseg.viewer.widget.segmentation import widget_simple_dt_ws
 from plantseg.viewer.widget.segmentation import widget_fix_over_under_segmentation_from_nuclei
 from plantseg.viewer.widget.segmentation import widget_lifted_multicut
-from plantseg.viewer.widget.proofreading.proofreading import widget_split_and_merge_from_scribbles
-from plantseg.viewer.widget.proofreading.proofreading import widget_clean_scribble
-from plantseg.viewer.widget.dataprocessing import widget_label_processing
-from PyQt5.QtCore import Qt
-import webbrowser
+from plantseg.viewer.widget.segmentation import widget_simple_dt_ws
 
 
 def setup_menu(container, path=None):
