+++ plantseg/viewer/widget/io.py
@@ -1,20 +1,19 @@
-from functools import partial
 from pathlib import Path
 from typing import List, Tuple
+from warnings import warn
 
 import numpy as np
 from magicgui import magicgui
 from napari.layers import Layer, Image, Labels
 from napari.types import LayerDataTuple
-from warnings import warn
+from napari.utils.notifications import show_info
 
-from plantseg.dataprocessing.functional.dataprocessing import image_rescale, compute_scaling_factor
 from plantseg.dataprocessing.functional.dataprocessing import fix_input_shape, normalize_01
+from plantseg.dataprocessing.functional.dataprocessing import image_rescale, compute_scaling_factor
 from plantseg.io import H5_EXTENSIONS, TIFF_EXTENSIONS, allowed_data_format
 from plantseg.io.io import load_tiff, load_h5, create_tiff
 from plantseg.viewer.dag_handler import dag_manager
 from plantseg.viewer.widget.utils import layer_properties
-from napari.utils.notifications import show_info
 
 
 def _check_layout_string(layout):
